Adapt old code to clean architecture
