--- app/src/main/java/com/team4/BoulderBuild/model/domain/Gym.kt
@@ -1,10 +1,0 @@
-package com.team4.BoulderBuild.model.domain
-
-data class Gym (
-    var id: Int?,
-    val name: String?,
-    val description: String?,
-    val lat : Double?,
-    val lon : Double?,
-    val imgSrc : String?
-)
