--- data/src/main/java/com/team4/data/MyClass.kt
@@ -1,3 +1,0 @@
-package com.team4.data
-
-class MyClass
