+++ data/src/main/java/com/team4/data/source/GymsLocalDataSource.kt
@@ -1,0 +1,11 @@
+package com.team4.data.source
+
+import com.team4.domain.Gym
+
+interface GymsLocalDataSource {
+    suspend fun saveGyms(gyms : List<Gym>)
+    suspend fun getAllGyms() : List<Gym>
+    suspend fun findGymById(id: Int) : Gym?
+    suspend fun update(gym: Gym)
+    suspend fun gymCount() : Int
+}
