+++ domain/src/main/java/com/team4/domain/Gym.kt
@@ -1,0 +1,10 @@
+package com.team4.domain
+
+data class Gym (
+    val id: Int,
+    val name: String,
+    val description: String,
+    val lat : Double,
+    val lon : Double,
+    val imgSrc : String
+)
