--- usecases/src/main/java/com/team4/lib/MyClass.kt
@@ -1,3 +1,0 @@
-package com.team4.lib
-
-class MyClass
