+++ usecases/src/main/java/com/team4/usecases/MyClass.kt
@@ -1,0 +1,3 @@
+package com.team4.usecases
+
+class MyClass
